Time series incoming

+++ settings.py
@@ -8,6 +8,6 @@ Created on Mon Dec  9 15:25:12 2019
 TRACK_WORDS = 'instagram'
 TABLE_NAME = "instagram"
 TABLE_ATTRIBUTES = "id_str VARCHAR(255), created_at DATETIME, text VARCHAR(255), \
-            polarity INT, subjectivity INT, user_created_at VARCHAR(255), user_location VARCHAR(255), \
+            polarity FLOAT, subjectivity INT, user_created_at VARCHAR(255), user_location VARCHAR(255), \
             user_description VARCHAR(255), user_followers_count INT, longitude DOUBLE, latitude DOUBLE, \
             retweet_count INT, favorite_count INT"
